Rollback to previous log compression↲
I had misunderstood how log compression is done in Tactron2. The old
code already did clipping same way as mentioned in Tacotron2. Althrough
the threshold is different between ours (1e-5) and Tacotron2 (0.01),
I think it's safe to keep the value as is because it has been tested
for months. We can try 0.01 later.

should fix #43